wire: add StatusBarFilePickerSpike target to Package.swift + consolidate into single-file spike
